+++ src/test/token-test.ts
@@ -10,7 +10,8 @@
  */
 
 import {expect} from 'chai';
-import {Token} from '../shady-css/token';
+
+import {Token, TokenType} from '../shady-css/token';
 
 describe('Token', () => {
   it('supports bitfield type comparison', () => {
@@ -19,6 +20,6 @@ describe('Token', () => {
     expect(token.is(4)).to.not.be.ok;
     expect(token.is(32 | 2)).to.be.ok;
     expect(token.is(4 | 64)).to.not.be.ok;
-    expect(token.is(128)).to.be.ok;
+    expect(token.is(128 as TokenType)).to.be.ok;
   });
 });